Refactor product catalog and category handling
Reworked product catalog logic to use Int64 for article IDs, improved category selection and display, and added GUI_ProductCatalogCategorieSelected for category-specific views. Updated hardware classes and related functions to support Int64 IDs and enhanced manufacturer/category filtering. Added test data initialization and fixed various UI and logic inconsistencies.

+++ VSCode/Elektrogrrosshandel/Elektrogrrosshandel/GUI/GUI_Menus/GUI_ProductCatalogArticelOverview.cs
@@ -24,7 +24,7 @@ namespace Elektrogrosshandel.GUI.GUI_Menus
         private static Panel PanelArtikelInfo()
         {
             List<Markup> articelInfo = new List<Markup>();
-            articelInfo = ComputerHardware.GetArticelInfoByID(int.Parse(ArticelID));
+            articelInfo = ComputerHardware.GetArticelInfoByID(Int64.Parse(ArticelID));
 
             if (articelInfo.Count == 0)
             {
